Add JavaDoc

+++ soundscribe-be/src/main/java/com/soundscribe/jvamp/JvampService.java
@@ -7,6 +7,9 @@ import java.io.File;
 import static com.soundscribe.jvamp.JvampFunctions.NOTES;
 import static com.soundscribe.jvamp.JvampFunctions.SMOOTHED_PITCH_TRACK;
 
+/**
+ * Executes jVamp plugins.
+ */
 public class JvampService {
     private Host host;
 
@@ -16,13 +19,23 @@ public class JvampService {
         host = new Host();
     }
 
+    /**
+     *
+     * @param fileWav Wav file for analysis.
+     * @param deleteAfter Boolean, if set on true wav file will be deleted after the process is completed.
+     */
     public void pyinSmoothedPitchTrack(File fileWav, boolean deleteAfter) {
         host.start(SMOOTHED_PITCH_TRACK, fileWav);
-        fileWav.delete();
+        if(deleteAfter) fileWav.delete();
     }
 
+    /**
+     * Generetes xmlFile with timestamp,length and value of notes in song.
+     * @param fileWav Wav file for analysis.
+     * @param deleteAfter Boolean, if set on true wav file will be deleted after the process is completed.
+     */
     public void pyinNotes(File fileWav, boolean deleteAfter) {
         host.start(NOTES, fileWav);
-        fileWav.delete();
+        if(deleteAfter) fileWav.delete();
     }
 }
